use one file per guess_os_stack_limit implementation
