+++ src/guess_os_stack_limit/fallback.rs
@@ -1,0 +1,4 @@
+#[inline(always)]
+pub unsafe fn guess_os_stack_limit() -> Option<usize> {
+    None
+}
